Added Lua methods: - EntityPlayer:GetMovingBoxContents - Level:GetMyosotisPickups
