+++ repentogon/LuaInterfaces/LuaLevel.cpp
@@ -3,6 +3,7 @@
 #include "HookSystem.h"
 
 #include "Level.h"
+#include "LuaEntitySaveState.h"
 
 LevelASM levelASM;
 
@@ -82,6 +83,13 @@ LUA_FUNCTION(Lua_SetForceSpecialQuest) {
 	return 0;
 }
 
+LUA_FUNCTION(Lua_GetMyosotisPickups) {
+	Game* level = lua::GetUserdata<Game*>(L, 1, lua::Metatables::LEVEL, "Level");
+	Lua_EntitiesSaveStateVector* ud = lua::place<Lua_EntitiesSaveStateVector>(L, lua::metatables::EntitiesSaveStateVectorMT);
+	ud->data = (&level->_myosotisPickups);
+	return 1;
+}
+
 LUA_FUNCTION(Lua_LevelIsAltPath) {
 	bool ret = false;
 	// If ForceSpecialQuest is -1, quest doors are disabled
@@ -154,6 +162,7 @@ HOOK_METHOD(LuaEngine, RegisterClasses, () -> void) {
 		{ "GetDimension", Lua_GetDimension},
 		{ "GetForceSpecialQuest", Lua_GetForceSpecialQuest },
 		{ "SetForceSpecialQuest", Lua_SetForceSpecialQuest },
+		{ "GetMyosotisPickups", Lua_GetMyosotisPickups },
 
 		{ "SetGreedWavesClearedWithoutRedHeartDamage", lua_LevelSetGreedWavesClearedWithoutRedHeartDamage },
 		{ "GetGreedWavesClearedWithoutRedHeartDamage", lua_LevelGetGreedWavesClearedWithoutRedHeartDamage },
